Tests: re-organize test directory structure
* Move all existing test file in a `Unit` subdirectory as these are unit tests.
    This allows for more easily adding a second set of (integration) tests at a later point in time.
* Let the subdirectory structure of the `tests\Unit` directory reflect the structure of the `src` directory.
* Let the class names of the test files reflect the class name + method name of the method under test.
